// samples/FluentSample/src/generated/models/DEVES_BusinessCentralModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface TimeRegistration {
  "@odata.etag"?: string;
  id: string;
  employeeNo: string;
  workLocation: string;
  workRoleCode: string;
  systemDateEntry: string;
  systemTimeEntry: string;
  systemDateExit: string;
  systemTimeExit: string;
  userDateEntry: string;
  userTimeEntry: string;
  userDateExit: string;
  userTimeExit: string;
  noOfHours: number;
  status: string;
  entryStatus: string;
  leavingStatus: string;
  entryMethod: string;
  entryEmployeeComment: string;
  originLogon: string;
  checkInDateTime: string;
  checkOutDateTime: string;
  durationMinutes: number;
}

export interface TimeRegistrationResponse {
  "@odata.context"?: string;
  value: TimeRegistration[];
}